Matrix indexing operator returns const ref instead of value
Even though the matrices in the tests were of a primitive type which are
cheap to copy, better performance can be achieved when the indexing
operator returns the value by reference, also for const objects.

One might think that it is faster to copy an int16_t value than an int16_t&
reference (pointers are 64 bit on current CPU-s), but I achieved 2x
speed-up in the regular multiplication algorithm when the const indexing
function returned const& instead of value.

The benchmarks do not show such a significant change for the FMMI
multiplication algorithm, though.