Add matrix implementation with runtime size: dmatrix
This implementation is more flexible regarding that the size of the
matrix can be defined with any expression calculated at run-time.
(The size has to be given as a constructor argument rather than a
template argument.)

This removes some of the benefits of the zero-cost abstractions
like expanding recursive calls by the constructor or using
constexpr if statements.

However, the binaries will be smaller, compilation will be faster.